requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.3.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
loguru==0.7.3
lxml==5.4.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
psycopg==3.2.6
psycopg-binary==3.2.6
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
